+++ src/main/scala/de/sciss/synth/Completion.scala
@@ -2,7 +2,7 @@
  *  Completion.scala
  *  (ScalaCollider)
  *
- *  Copyright (c) 2008-2016 Hanns Holger Rutz. All rights reserved.
+ *  Copyright (c) 2008-2018 Hanns Holger Rutz. All rights reserved.
  *
  *  This software is published under the GNU Lesser General Public License v2.1+
  *
@@ -14,14 +14,15 @@
 package de.sciss.synth
 
 import de.sciss.osc.Packet
-import language.implicitConversions
+
+import scala.language.implicitConversions
 
 object Completion {
-  implicit def fromPacket[T](p: Packet): Completion[T]        = Completion[T](Some((_: T) => p), scala.None)
-  implicit def fromFunction[T](fun: T => Unit): Completion[T] = Completion[T](scala.None, Some(fun))
+  implicit def fromPacket  [A](p: Packet     ): Completion[A] = Completion[A](Some((_: A) => p), scala.None)
+  implicit def fromFunction[A](fun: A => Unit): Completion[A] = Completion[A](scala.None, Some(fun))
 
   val None: Completion[Any] = Completion(scala.None, scala.None)
 }
-final case class Completion[-T](message: Option[T => Packet], action: Option[T => Unit]) {
-  def mapMessage(t: T): Option[Packet] = message.map(_.apply(t))
+final case class Completion[-A](message: Option[A => Packet], action: Option[A => Unit]) {
+  def mapMessage(t: A): Option[Packet] = message.map(_.apply(t))
 }