+++ src/main/scala/de/sciss/synth/Group.scala
@@ -2,7 +2,7 @@
  *  Group.scala
  *  (ScalaCollider)
  *
- *  Copyright (c) 2008-2016 Hanns Holger Rutz. All rights reserved.
+ *  Copyright (c) 2008-2018 Hanns Holger Rutz. All rights reserved.
  *
  *  This software is published under the GNU Lesser General Public License v2.1+
  *
@@ -14,17 +14,13 @@
 package de.sciss.synth
 
 object Group {
-  def apply(server: Server): Group = apply(server, server.nextNodeID())
+  def apply(server: Server): Group = apply(server, server.nextNodeId())
 
   def apply(): Group = apply(Server.default)
 }
 
 final case class Group(server: Server, id: Int)
   extends Node {
-
-//  def this(server: Server) = this(server, server.nextNodeID())
-//
-//  def this() = this(Server.default)
 
   def newMsg(target: Node, addAction: AddAction) =
     message.GroupNew(message.GroupNew.Data(id, addAction.id, target.id))