+++ src/test/scala/de/sciss/synth/Issue49.scala
@@ -1,0 +1,14 @@
+package de.sciss.synth
+
+trait Issue49 {
+  import ugen._
+  import Ops._
+
+  def s: Server = Server.default
+
+  play { SinOsc.ar }  // ok
+  playWith(target = s, addAction = addToTail) { SinOsc.ar } // ok
+
+  playWith(s) { SinOsc.ar } // ok
+  playWith(target = s) { SinOsc.ar } // ok
+}